Merged Betaflight code (#16)
* Use getRSSI() for telemetry detection

* FF max rate limit and jitter reduction

* Reorder and rename simplified pids

Order and names are now the same as in the configurator

* Update MSP protocol

* Disable reboot for dynamic idle change

* Fix string writes to board info table

* Update simplified_tuning.lua

* Update firmware requirements

* Clear LCD for CMS script at startup

Clears the LCD and draws the "refresh[whateverbuttonrefreshis]" at the top row when the script is launched.

The current behaviour is to not clear the screen when the script is launched. If it's not responding and a manual refresh is required, there's no way for the user to know this. By clearing the screen and showing the message, the user should get an idea of what has to be done.

* CMS periodic refresh when stick movement detected

* Make BOARD_INFO download conditional to apiVersion >=1.043

* CMS move initial screen.clear() to run function

* CMS fixes

* Rename script in TOOLS menu

Rename to "Betaflight Config" to match the name in the title bar. Also fixed the downloading vtx tables string.

* Back to Betaflight setup

"config" makes it one character too long to display on 128x64.

* Update readme for 1.6.0

* Add BF lua logo to readme

* Ghost MSP support

* Ignore .vscode

add .vscode to .gitignore

* Add motor output limit

* Add Profiles page.

* fix shebang in bin/build.sh

Using env is a more portable approach.

* Auto generation of release notes.

* Moving to github actions, to attach on release.

* Read lua input buffer until empty

* Remove gyro/pid rate for BF 4.2

Co-authored-by: Hans Christian Olaussen <41271048+klutvott123@users.noreply.github.com>
Co-authored-by: haslinghuis <mark@numloq.nl>
Co-authored-by: Štěpán Dalecký <daleckystepan@gmail.com>
Co-authored-by: Richard Marko <srk@48.io>
Co-authored-by: blckmn <blackman@xtra.com.au>
Co-authored-by: J Blackman <blckmn@users.noreply.github.com>